# Webp encoding: Reduce allocations.