--- a/src/ImageSharp/Formats/Webp/Lossless/HuffmanUtils.cs
+++ b/src/ImageSharp/Formats/Webp/Lossless/HuffmanUtils.cs
@@ -203,10 +203,10 @@ namespace SixLabors.ImageSharp.Formats.Webp.Lossless
 
                 // Build the Huffman tree.
 #if NET5_0_OR_GREATER
-                Span<HuffmanTree> treeSlice = tree.AsSpan().Slice(0, treeSize);
+                Span<HuffmanTree> treeSlice = tree.AsSpan(0, treeSize);
                 treeSlice.Sort(HuffmanTree.Compare);
 #else
-                HuffmanTree[] treeCopy = tree.AsSpan().Slice(0, treeSize).ToArray();
+                HuffmanTree[] treeCopy = tree.AsSpan(0, treeSize).ToArray();
                 Array.Sort(treeCopy, HuffmanTree.Compare);
                 treeCopy.AsSpan().CopyTo(tree);
 #endif