… src/ImageSharp/Formats/Webp/Lossless/Vp8LEncoder.cs
… src/ImageSharp/Formats/Webp/Lossless/Vp8LHistogram.cs
… src/ImageSharp/Formats/Webp/Lossless/Vp8LStreaks.cs
… src/ImageSharp/Formats/Webp/Lossless/WebpLosslessDecoder.cs
… src/ImageSharp/Formats/Webp/Lossy/LossyUtils.cs
… src/ImageSharp/Formats/Webp/Lossy/QuantEnc.cs
… src/ImageSharp/Formats/Webp/Lossy/Vp8EncIterator.cs
… src/ImageSharp/Formats/Webp/Lossy/Vp8Encoder.cs
… src/ImageSharp/Formats/Webp/Lossy/Vp8Encoding.cs
… src/ImageSharp/Formats/Webp/Lossy/Vp8Histogram.cs
… src/ImageSharp/Formats/Webp/Lossy/Vp8ModeScore.cs
… src/ImageSharp/Formats/Webp/Lossy/Vp8Residual.cs
… src/ImageSharp/Formats/Webp/Lossy/WebpLossyDecoder.cs
… src/ImageSharp/Formats/Webp/WebpCommonUtils.cs
… tests/ImageSharp.Tests/Formats/WebP/PredictorEncoderTests.cs