### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description

This PR is meant to reduce the allocations during webp encoding significantly. 

Related to #1786

Before:

```
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19043.1288 (21H1/May2021Update)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.100-rc.2.21505.57
  [Host]     : .NET 5.0.11 (5.0.1121.47308), X64 RyuJIT
  Job-OKVXPJ : .NET 5.0.11 (5.0.1121.47308), X64 RyuJIT
  Job-PBVPYV : .NET Core 3.1.20 (CoreCLR 4.700.21.47003, CoreFX 4.700.21.47101), X64 RyuJIT
  Job-SLAEGF : .NET Framework 4.8 (4.8.4420.0), X64 RyuJIT

IterationCount=3  LaunchCount=1  WarmupCount=3
|                     Method |        Job |              Runtime |             Arguments |    TestImage |      Mean |      Error |    StdDev | Ratio | RatioSD |       Gen 0 |     Gen 1 |     Gen 2 |  Allocated |
|--------------------------- |----------- |--------------------- |---------------------- |------------- |----------:|-----------:|----------:|------:|--------:|------------:|----------:|----------:|-----------:|
|        'Magick Webp Lossy' | Job-OKVXPJ |             .NET 5.0 | /p:DebugType=portable | Png/Bike.png |  26.78 ms |   4.110 ms |  0.225 ms |  0.15 |    0.00 |           - |         - |         - |      68 KB |
|    'ImageSharp Webp Lossy' | Job-OKVXPJ |             .NET 5.0 | /p:DebugType=portable | Png/Bike.png | 285.95 ms |  33.692 ms |  1.847 ms |  1.56 |    0.03 | 135000.0000 |         - |         - | 552,714 KB |
|     'Magick Webp Lossless' | Job-OKVXPJ |             .NET 5.0 | /p:DebugType=portable | Png/Bike.png | 183.18 ms |  34.303 ms |  1.880 ms |  1.00 |    0.00 |           - |         - |         - |     520 KB |
| 'ImageSharp Webp Lossless' | Job-OKVXPJ |             .NET 5.0 | /p:DebugType=portable | Png/Bike.png | 347.68 ms |  67.105 ms |  3.678 ms |  1.90 |    0.02 |  34000.0000 | 5000.0000 | 2000.0000 | 161,675 KB |
|                            |            |                      |                       |              |           |            |           |       |         |             |           |           |            |
|        'Magick Webp Lossy' | Job-PBVPYV |        .NET Core 3.1 |               Default | Png/Bike.png |  26.77 ms |   8.193 ms |  0.449 ms |  0.15 |    0.00 |           - |         - |         - |      68 KB |
|    'ImageSharp Webp Lossy' | Job-PBVPYV |        .NET Core 3.1 |               Default | Png/Bike.png | 294.26 ms | 156.747 ms |  8.592 ms |  1.62 |    0.08 | 135000.0000 |         - |         - | 552,713 KB |
|     'Magick Webp Lossless' | Job-PBVPYV |        .NET Core 3.1 |               Default | Png/Bike.png | 181.80 ms |  69.170 ms |  3.791 ms |  1.00 |    0.00 |           - |         - |         - |     520 KB |
| 'ImageSharp Webp Lossless' | Job-PBVPYV |        .NET Core 3.1 |               Default | Png/Bike.png | 357.09 ms | 433.254 ms | 23.748 ms |  1.97 |    0.16 |  34000.0000 | 5000.0000 | 2000.0000 | 161,668 KB |
|                            |            |                      |                       |              |           |            |           |       |         |             |           |           |            |
|        'Magick Webp Lossy' | Job-SLAEGF | .NET Framework 4.7.2 |               Default | Png/Bike.png |  26.63 ms |   2.530 ms |  0.139 ms |  0.14 |    0.00 |           - |         - |         - |      68 KB |
|    'ImageSharp Webp Lossy' | Job-SLAEGF | .NET Framework 4.7.2 |               Default | Png/Bike.png | 429.41 ms |  38.772 ms |  2.125 ms |  2.33 |    0.01 | 135000.0000 |         - |         - | 554,351 KB |
|     'Magick Webp Lossless' | Job-SLAEGF | .NET Framework 4.7.2 |               Default | Png/Bike.png | 183.96 ms |  33.354 ms |  1.828 ms |  1.00 |    0.00 |           - |         - |         - |     523 KB |
| 'ImageSharp Webp Lossless' | Job-SLAEGF | .NET Framework 4.7.2 |               Default | Png/Bike.png | 428.93 ms |  60.292 ms |  3.305 ms |  2.33 |    0.04 |  34000.0000 | 5000.0000 | 2000.0000 | 162,125 KB |
```

After

```
|                     Method |        Job |              Runtime |    TestImage |      Mean |      Error |    StdDev | Ratio | RatioSD |     Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|--------------------------- |----------- |--------------------- |------------- |----------:|-----------:|----------:|------:|--------:|----------:|----------:|----------:|----------:|
|        'Magick Webp Lossy' | Job-TRNEBU |             .NET 5.0 | Png/Bike.png |  26.38 ms |   4.031 ms |  0.221 ms |  0.14 |    0.00 |         - |         - |         - |     68 KB |
|    'ImageSharp Webp Lossy' | Job-TRNEBU |             .NET 5.0 | Png/Bike.png | 232.16 ms |  50.122 ms |  2.747 ms |  1.27 |    0.02 | 4000.0000 |         - |         - | 16,704 KB |
|     'Magick Webp Lossless' | Job-TRNEBU |             .NET 5.0 | Png/Bike.png | 182.79 ms |   8.574 ms |  0.470 ms |  1.00 |    0.00 |         - |         - |         - |    520 KB |
| 'ImageSharp Webp Lossless' | Job-TRNEBU |             .NET 5.0 | Png/Bike.png | 350.50 ms |  66.007 ms |  3.618 ms |  1.92 |    0.02 | 8000.0000 | 3000.0000 | 1000.0000 | 61,556 KB |
|                            |            |                      |              |           |            |           |       |         |           |           |           |           |
|        'Magick Webp Lossy' | Job-XLXDYZ |        .NET Core 3.1 | Png/Bike.png |  25.11 ms |   8.152 ms |  0.447 ms |  0.14 |    0.01 |         - |         - |         - |     68 KB |
|    'ImageSharp Webp Lossy' | Job-XLXDYZ |        .NET Core 3.1 | Png/Bike.png | 241.75 ms | 245.192 ms | 13.440 ms |  1.36 |    0.11 | 4000.0000 |         - |         - | 16,703 KB |
|     'Magick Webp Lossless' | Job-XLXDYZ |        .NET Core 3.1 | Png/Bike.png | 178.20 ms | 102.407 ms |  5.613 ms |  1.00 |    0.00 |         - |         - |         - |    520 KB |
| 'ImageSharp Webp Lossless' | Job-XLXDYZ |        .NET Core 3.1 | Png/Bike.png | 332.18 ms | 112.166 ms |  6.148 ms |  1.87 |    0.07 | 8000.0000 | 3000.0000 | 1000.0000 | 61,555 KB |
|                            |            |                      |              |           |            |           |       |         |           |           |           |           |
|        'Magick Webp Lossy' | Job-XXCPPX | .NET Framework 4.7.2 | Png/Bike.png |  26.47 ms |   1.311 ms |  0.072 ms |  0.14 |    0.00 |         - |         - |         - |     68 KB |
|    'ImageSharp Webp Lossy' | Job-XXCPPX | .NET Framework 4.7.2 | Png/Bike.png | 473.98 ms |  87.433 ms |  4.793 ms |  2.59 |    0.02 | 4000.0000 |         - |         - | 16,753 KB |
|     'Magick Webp Lossless' | Job-XXCPPX | .NET Framework 4.7.2 | Png/Bike.png | 182.86 ms |  26.630 ms |  1.460 ms |  1.00 |    0.00 |         - |         - |         - |    523 KB |
| 'ImageSharp Webp Lossless' | Job-XXCPPX | .NET Framework 4.7.2 | Png/Bike.png | 518.88 ms | 128.756 ms |  7.058 ms |  2.84 |    0.05 | 8000.0000 | 3000.0000 | 1000.0000 | 61,697 KB |
```