@brianpopow 2021-11-07
@JimBobSquarePants any objections that we merge this PR as it is? I really would like to see how it does in combination with the other optimizations. I think the reduced allocations can be beneficial to performance, too.